fix cosimulation example
