+++ examples/cosimulation.v.etc/Makefile.conf
@@ -17,7 +17,7 @@ CXXLDFLAGS := $(blackbox_objdir)/verilated.o $(blackbox_objdir)/Vblackbox__ALL.a
 
 cosimulation.opt: $(blackbox_objdir)/Vblackbox__ALL.a $(blackbox_objdir)/verilated.o
 
-ifeq (, $(shell which verilator))
+ifeq (, $(shell command -v verilator))
 $(warning Verilator unavailable, skipping cosimulation test)
 pass:
 .DEFAULT_GOAL := pass
