+++ examples/cosimulation.v.etc/cosimulation.cpp
@@ -1,33 +1,7 @@
 /*! Custom Cuttlesim driver that implements a Kôika extfun using a Verilator model !*/
 #include "cosimulation.hpp"
-#include "blackbox.obj_dir.opt/Vblackbox.h"
 #include <verilated.h>
 
-using Dut = Vblackbox;
-
-class BlackboxWrapper {
-public:
-  Dut dut;
-
-  void cycle() {
-    dut.CLK = 1;
-    dut.eval();
-    dut.CLK = 0;
-    dut.eval();
-  }
-
-  BlackboxWrapper() : dut{} {}
-};
-
-struct extfuns {
-  BlackboxWrapper blackbox_driver;
-
-  bits<32> blackbox(bits<32> input) {
-    blackbox_driver.dut.in = input.v;
-    blackbox_driver.cycle();
-    return  bits<32>{blackbox_driver.dut.out};
-  }
-};
 
 class simulator final : public module_cosimulation<extfuns> {
   void strobe() const {
